+++ src/server/config.ts
@@ -443,6 +443,9 @@ export const STATS = {
     errors: {
         internal: "errors.internal",
     },
+    request: {
+        duration: "request.duration",
+    },
 }
 
 export function ensureProtocol(